requirements.txt
line-bot-sdk
flask
gunicorn
Pillow
cloudinary
